Conflicts:
	NMSSH.podspec
	NMSSH.xcodeproj/project.pbxproj
	NMSSH/NMSFTPFile.m

+++ NMSSH/NMSSHChannel.m
@@ -1,5 +1,5 @@
 #import "NMSSHChannel.h"
-#import "socket_helper.h"
+#import "NMSSH+Protected.h"
 
 @interface NMSSHChannel ()
 @property (nonatomic, strong) NMSSHSession *session;
@@ -618,7 +618,7 @@
                 return NO;
             }
             got += rc;
-            if (progress && !progress(got, (NSUInteger)fileinfo.st_size)) {
+            if (progress && !progress((NSUInteger)got, (NSUInteger)fileinfo.st_size)) {
                 close(localFile);
                 [self closeChannel];
                 return NO;
